requirements.txt
flask==3.1.0
python-dotenv==1.1.0
requests==2.32.3
python-telegram-bot==13.15
pymysql==1.1.1
flask-sqlalchemy==3.1.1
alembic==1.16.1